# apps/api/src/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Compasso — API da Luna
  version: 1.0.0
  summary: Agenda, aulas e tarefas do Compasso para a assistente de voz Luna.
  description: |
    Porta servidor a servidor para a Luna consultar e criar itens da agenda por voz
    (ADR-0012). Cada chamada acontece no meio de uma resposta falada: previsibilidade vem antes
    de recursos.

    **Contrato de tempo**
    - Todo instante, na entrada e na saída, é ISO 8601 **com deslocamento**
      (`2026-09-26T14:00:00-03:00`). Sem deslocamento → `400 validation_error`.
    - O fuso de referência é `America/Sao_Paulo`; toda saída vem nele.
    - Dia inteiro usa só a data (`"2026-09-26"`) com `all_day: true`, nunca meia-noite UTC. Em
      evento de dia inteiro, `end` é o **último dia, inclusive**.
    - A API não interpreta linguagem natural ("amanhã", "sexta"): a Luna resolve a data antes.
    - Recorrências já vêm expandidas: cada ocorrência é um item, com o `id` da série,
      `recurring: true` e `occurrence_date`. O mesmo `id` pode aparecer várias vezes numa lista.

    **Erros**: sempre `{"error":{"code","message","field?"}}`. Decida pelo `code` (estável); a
    `message` é curta, em português, pronta para ser falada.

    **Limites**: 120 requisições por minuto por token (`429 rate_limited` com `Retry-After`);
    corpo até 16 KiB; intervalo da agenda até 366 dias; página padrão de 100 itens (máx. 200).

    **Não suportado nesta versão**: editar, mover, remanejar ou apagar evento ou tarefa; criar
...
servers:
...
security:
  - bearer: []
...
paths:
...
  /agenda:
    get:
      tags: [agenda]
      operationId: getAgenda
      summary: Eventos, aulas e tarefas de um intervalo, numa chamada
      description: |
        Tudo o que cai em `[from, to)`, ordenado por início: eventos (com recorrências
        expandidas), aulas da grade (inclusive canceladas, com `cancelled: true`) e tarefas com
        prazo no intervalo. Tarefa sem prazo não tem lugar na agenda: use `GET /tasks`.

        Ordem: instante de início (dia inteiro conta como 00:00 do dia), dia inteiro antes, depois
        evento, aula, tarefa, e por fim o título.
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date-time }
          example: '2026-09-26T00:00:00-03:00'
        - name: to
          in: query
          required: true
          description: Exclusivo. Até 366 dias depois de `from`.
          schema: { type: string, format: date-time }
          example: '2026-09-27T00:00:00-03:00'
        - name: types
          in: query
          description: Subconjunto de `event,class,task`, separado por vírgula. Padrão, todos.
          schema: { type: string }
          example: event,class
        - name: q
          in: query
          description: |
            Busca por título (e nome da disciplina, nas aulas), sem diferenciar acento nem caixa.
            Casa quando toda palavra de 3 letras ou mais aparece no título: "prova de cálculo"
            encontra "Prova de Cálculo II". Combine com um intervalo largo para perguntas como
            "quando é a prova?".
          schema: { type: string, minLength: 1, maxLength: 100 }
          example: prova de cálculo
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Itens do intervalo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListaDeItens' }
              examples:
                amanha:
                  summary: GET /agenda?from=2026-09-26T00:00:00-03:00&to=2026-09-27T00:00:00-03:00
                  value:
                    timezone: America/Sao_Paulo
                    items:
                      - id: 01a0d9ec-865a-755d-825c-1a30799cd8e7
                        type: event
                        title: Aniversário da Ana
                        start: '2026-09-26'
                        end: '2026-09-26'
                        all_day: true
                        location: null
                        recurring: false
                        occurrence_date: null
                      - id: 01a0d9ec-864c-722c-8c33-5754b2a23524
                        type: event
                        title: Treino na academia
                        start: '2026-09-26T07:00:00-03:00'
                        end: '2026-09-26T08:00:00-03:00'
                        all_day: false
                        location: null
                        recurring: true
                        occurrence_date: '2026-09-26'
                      - id: 01a0d9ec-8643-723b-bfb9-03d4e79dc466
                        type: class
                        title: Cálculo II
                        start: '2026-09-26T08:00:00-03:00'
                        end: '2026-09-26T10:00:00-03:00'
                        all_day: false
                        location: B-204
                        recurring: true
                        occurrence_date: '2026-09-26'
                        subject: Cálculo II
                        professor: Marcos Lima
                        cancelled: false
                      - id: 01a0d9ec-8657-7295-b5a6-3c1813e873fa
                        type: event
                        title: Dentista
                        start: '2026-09-26T14:00:00-03:00'
                        end: '2026-09-26T15:00:00-03:00'
                        all_day: false
                        location: null
                        recurring: false
                        occurrence_date: null
                      - id: 01a0d9ec-8674-75ec-b821-046086939197
                        type: task
                        title: Entregar relatório de física
                        start: '2026-09-26T23:59:00-03:00'
                        end: null
                        all_day: false
                        location: null
                        recurring: false
                        occurrence_date: null
                        due: '2026-09-26T23:59:00-03:00'
                        done: false
                    next_cursor: null
                busca:
                  summary: GET /agenda?from=2026-09-25T00:00:00-03:00&to=2026-11-24T00:00:00-03:00&types=event&q=prova%20de%20calculo
                  value:
                    timezone: America/Sao_Paulo
                    items:
                      - id: 01a0d9ec-8671-70f7-82c0-26475bd1c764
                        type: event
                        title: Prova de Cálculo II
                        start: '2026-10-09T08:00:00-03:00'
                        end: '2026-10-09T10:00:00-03:00'
                        all_day: false
                        location: null
                        recurring: false
                        occurrence_date: null
                    next_cursor: null
        '400':
          description: Parâmetro inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erro' }
              example:
                error:
                  code: validation_error
                  message: from precisa ser data e hora com fuso, como 2026-09-26T00:00:00-03:00
                  field: from
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
...
  /tasks/{id}/complete:
    post:
      tags: [tarefas]
      operationId: completeTask
      summary: Marcar tarefa como feita
      description: |
        Credita XP e moedas como no app. Repetir é seguro: concluir a concluída não credita de
        novo e devolve `done: true`. Em tarefa recorrente, `occurrence_date` é obrigatório (a data
        da ocorrência, como veio na agenda); em tarefa simples, proibido. Não há desfazer por
        aqui: é no app.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                occurrence_date: { type: string, format: date }
            example: { occurrence_date: '2026-09-26' }
      responses:
        '200':
          description: A tarefa (ou a ocorrência) como ficou.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tarefa' }
              example:
                id: 01a0d9ec-8677-7501-a9bd-cf918809bd77
                type: task
                title: Comprar ração do gato
                start: null
                end: null
                all_day: false
                location: null
                recurring: false
                occurrence_date: null
                due: null
                done: true
        '400':
          description: Falta ou sobra `occurrence_date`, ou a tarefa não acontece nessa data.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erro' }
              example:
                error:
                  code: validation_error
                  message: 'essa tarefa se repete: informe occurrence_date'
                  field: occurrence_date
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Não existe, foi excluída ou não é tarefa (evento não se conclui por aqui).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Erro' }
              example: { error: { code: not_found, message: não encontrei essa tarefa } }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        Token de serviço gerado no app (Perfil → Configurações → Luna), mostrado uma vez.
        Escopos: `agenda:read` (consultas) e `agenda:write` (criar e concluir). Revogado ou
        errado → 401; válido sem o escopo da rota → 403. O token de serviço não acessa nenhuma
        rota fora de /api/v1. Trocar a senha no app revoga todos os tokens, inclusive o da Luna.
  parameters:
    limit:
      name: limit
      in: query
      description: Itens por página, de 1 a 200. Um dia ou uma semana cabem na página padrão.
      schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
    cursor:
      name: cursor
      in: query
      description: O `next_cursor` da resposta anterior, com os mesmos filtros.
      schema: { type: string }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        1 a 200 caracteres ASCII visíveis, único por intenção de criação (ex.: id do turno de
        voz). Reenviar o mesmo pedido com a mesma chave devolve o item já gravado com 200 e
        `Idempotent-Replayed: true`, sem duplicar, mesmo depois de reinício da API. A mesma chave
        com outro pedido → 409 `conflict`. Sem a chave, cada POST de criação cria um item.
      schema: { type: string, minLength: 1, maxLength: 200 }
      example: luna-turno-8f3a1c
  schemas:
...
    DataOuInstante:
      description: Só a data quando `all_day`; senão data e hora com deslocamento de São Paulo.
      anyOf:
        - { type: string, format: date, examples: ['2026-09-26'] }
        - { type: string, format: date-time, examples: ['2026-09-26T14:00:00-03:00'] }
...
    Evento:
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          required: [type]
          properties:
            type: { const: event }
            end:
              description: |
                Dia inteiro: a data do último dia, inclusive. Com horário: o fim, ou `null` em
                evento sem duração.
...
    Tarefa:
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          required: [type, due, done]
          properties:
            type: { const: task }
            start: { description: Igual a `due` (para a agenda ordenar por um campo só). }
            end: { type: 'null' }
            due:
              anyOf: [{ $ref: '#/components/schemas/DataOuInstante' }, { type: 'null' }]
            done: { type: boolean }
    Item:
      oneOf:
        - $ref: '#/components/schemas/Evento'
        - $ref: '#/components/schemas/Aula'
        - $ref: '#/components/schemas/Tarefa'
      discriminator:
        propertyName: type
        mapping:
          event: '#/components/schemas/Evento'
          class: '#/components/schemas/Aula'
          task: '#/components/schemas/Tarefa'
    ListaDeItens:
      type: object
      required: [timezone, items, next_cursor]
      properties:
        timezone: { const: America/Sao_Paulo }
        items:
          type: array
          items: { $ref: '#/components/schemas/Item' }
        next_cursor:
          type: [string, 'null']
          description: '`null` quando não há mais páginas.'
...
    Erro:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                [
                  unauthorized,
                  forbidden,
                  not_found,
                  validation_error,
                  conflict,
                  rate_limited,
                  internal,
                ]
            message:
              type: string
              description: Curta, em português, pode ser falada.
            field:
              type: string
              description: Campo, parâmetro ou cabeçalho culpado, quando há um.
  responses:
...
    Unauthorized:
      description: Token ausente, errado ou revogado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
          example: { error: { code: unauthorized, message: token ausente ou inválido } }
    Forbidden:
      description: Token válido sem o escopo da rota.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
          example: { error: { code: forbidden, message: este token só pode consultar a agenda } }
...
    RateLimited:
      description: Mais de 120 requisições no último minuto com este token.
      headers:
        Retry-After:
          description: Segundos até poder tentar de novo.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
          example:
            error: { code: rate_limited, message: muitas requisições; tente de novo em 60 s }
    Internal:
      description: Falha real do Compasso (banco fora do ar, bug). Só aqui há 5xx.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
          example: { error: { code: internal, message: erro interno no Compasso } }